src/scripts/requirements.txt
neuprint-python
vfb-connect